Removing TODOs and updating twist)controller.py
Note
In the walkthrough, only 400 Nm of torque is applied to hold the vehicle This turns out to be slightly less than the amount of force needed, and Carla roll forward with only 400Nm of torque. To prevent Carla from moving you should apply approximately 700 Nm of torque. I updated the code to apply appropriate torque based on the  testing environment.